+++ app/src/main/java/com/mapzip/ppang/mapzipproject/ui/mapinfos/InfosListFragment.java
@@ -1,5 +1,6 @@
 package com.mapzip.ppang.mapzipproject.ui.mapinfos;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.app.Fragment;
@@ -7,6 +8,7 @@ import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
 import com.mapzip.ppang.mapzipproject.R;
 
@@ -45,8 +47,41 @@ public class InfosListFragment extends Fragment {
         View rootView = inflater.inflate(R.layout.fragment_infos_list, container, false);
 
         mRecyclerView = (RecyclerView) rootView.findViewById(R.id.infos_list_recyclerview);
-        mRecyclerView.setAdapter(new ReviewsAdapter(getActivity()));
+        mRecyclerView.setAdapter(new InfosAdapter(getActivity()));
 
         return rootView;
+    }
+
+    private static class InfosAdapter extends RecyclerView.Adapter {
+        private final static int LOCATION_ITEM_TYPE = 0;
+
+        public InfosAdapter(Context mContext) {
+        }
+
+        @Override
+        public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+            return null;
+        }
+
+        @Override
+        public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
+
+        }
+
+        @Override
+        public int getItemCount() {
+            return 0;
+        }
+    }
+
+    public class ViewHolder extends RecyclerView.ViewHolder{
+
+        public TextView nameText;
+        public TextView addressText;
+        public TextView reviewCountsText;
+
+        public ViewHolder(View itemView) {
+            super(itemView);
+        }
     }
 }
